requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.1
quickchart-io==2.0.0
requests==2.28.1
urllib3==1.26.12
